fix: Update package name extraction from codeartifact path
When we added support for multiple configuration sections in commit
46e83fe03a305da673b20e7e1da6a2a93c8b6b6d, we inadvertently broke the
path parsing logic. A typical error would look something similar to
this:

    too many values to unpack (expected 3)
    at ~/path/to/keyrings/codeartifact.py:151 in get_password
      150│ # Use the second and third parts as repository type and name.
    → 151│ _, repository_type, repository_name = parts

The core of the problem was that our code assumed that codeartifact
paths look like /pypi/repo-name, but this assumption is false. Normally,
what we see is actually /pypi/repo-name/simple/ as documented in the AWS
documentation.

We also moved away from using PurePath to parse our paths. PurePath is
OS-dependent and behaves differently based on the platform we currently
run our Python interpretea on, but this is now what we want when parsing
paths in urls. Since valid codeartifact paths have very little variety
in them, we oped to parse them using a relatively simple regular
expression.

Our regular expression for validating url paths is a bit less strict
than what is described in AWS docs. The codeartifact docs state that all
urls need to end with a /simple/ suffix or things will not work. But in
practice, paths without a trailing / also work fine as redirect to
/-suffixed path as required by PEP 503.