feat: option to assume role to get credentials
When fetching credentials after obtaining the boto3 session, we check if
the configuration has an optional `assume_role` field, which indicates
that role needs to be assumed to obtain the credentials.

If found, the role is assumed and used to obtain the credentails. If not
the original boto3 session is used to obtain the credentials.